refactor(math): split monolith into per-indicator submodules
The `math/__init__.py` monolith hosted all 31 math operators and
transforms in a single file. Each operator now lives in its own module
(`math/add.py`, `math/sub.py`, ...) matching the one-file-per-indicator
convention used by every other category. The package `__init__` keeps
re-exporting every name, so all existing import paths
(`from pandas_ta_classic.math import add`, `ta.math.add`, etc.) continue
to work unchanged.
