+++ pandas_ta_classic/math/acos.py
@@ -1,0 +1,22 @@
+from typing import Any, Optional
+
+import numpy as np
+from pandas import Series
+
+from pandas_ta_classic.utils import apply_fill, apply_offset, get_offset, verify_series
+
+
+def acos(
+    close: Series, offset: Optional[int] = None, **kwargs: Any
+) -> Optional[Series]:
+    """Vector Trigonometric ACos (TA-Lib: ACOS)."""
+    close = verify_series(close)
+    if close is None:
+        return None
+    offset = get_offset(offset)
+    result = close.apply(np.arccos)
+    result = apply_offset(result, offset)
+    result = apply_fill(result, **kwargs)
+    result.name = "ACOS"
+    result.category = "math"
+    return result
